using VertId = uint16_t;

+++ src/crab/splitdbm/definitions.hpp
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT
 #pragma once
 
+#include <cstdint>
 #include <unordered_set>
 
 #include "arith/num_big.hpp"
@@ -25,7 +26,7 @@ namespace splitdbm {
  **/
 using Weight = prevail::Number;
 
-using VertId = unsigned int;
+using VertId = uint16_t;
 using VertSet = std::unordered_set<VertId>;
 
 }
